+++ src/World.hpp
@@ -18,7 +18,7 @@ class World {
         // list of all collidable surfaces. Starting with pad(s)
         std::vector<Ground> grounds;
         // gravitational acceleration, meters/s^2
-        constexpr static const float g = 1.62;
+        constexpr static const double g = 1.62;
         static const int SEGMENT_WIDTH = 40;
         static const int MAX_DY = 40;
